feat: make profile channel article to ssr
