+++ src/app/(normal)/profile/(profile)/[source]/[id]/layout.tsx
@@ -5,6 +5,7 @@ import { ProfilePageLayout } from '@/app/(normal)/profile/pages/ProfilePageLayou
 import { SourceInURL } from '@/constants/enum.js';
 import { isProfilePageSource } from '@/helpers/isProfilePageSource.js';
 import { resolveSourceFromUrlNoFallback } from '@/helpers/resolveSource.js';
+import { setupLocaleForSSR } from '@/i18n/index.js';
 
 export default function Layout({
     children,
@@ -19,6 +20,8 @@ export default function Layout({
     const source = resolveSourceFromUrlNoFallback(params.source);
     if (!source || !isProfilePageSource(source)) notFound();
     const identity = { source, id };
+
+    setupLocaleForSSR();
 
     return <ProfilePageLayout identity={identity}>{children}</ProfilePageLayout>;
 }
