+++ src/components/Profile/Info.tsx
@@ -1,54 +1,29 @@
-import { plural, Trans } from '@lingui/macro';
-import { useInfiniteQuery } from '@tanstack/react-query';
+import { Plural, Trans } from '@lingui/macro';
 
 import { Avatar } from '@/components/Avatar.js';
-import { AvatarGroup } from '@/components/AvatarGroup.js';
 import { BioMarkup } from '@/components/Markup/BioMarkup.js';
+import { NoSSR } from '@/components/NoSSR.js';
+import { Mutuals } from '@/components/Profile/Mutuals.js';
 import { ProfileAction } from '@/components/Profile/ProfileAction.js';
 import { ProfileVerifyBadge } from '@/components/ProfileVerifyBadge/index.js';
 import { SocialSourceIcon } from '@/components/SocialSourceIcon.js';
 import { TextOverflowTooltip } from '@/components/TextOverflowTooltip.js';
 import { FollowCategory, Source } from '@/constants/enum.js';
-import { EMPTY_LIST } from '@/constants/index.js';
 import { Link } from '@/esm/Link.js';
 import { classNames } from '@/helpers/classNames.js';
 import { nFormatter } from '@/helpers/formatCommentCounts.js';
 import { getLargeTwitterAvatar } from '@/helpers/getLargeTwitterAvatar.js';
-import { isSameProfile } from '@/helpers/isSameProfile.js';
-import { narrowToSocialSource } from '@/helpers/narrowToSocialSource.js';
 import { resolveProfileUrl } from '@/helpers/resolveProfileUrl.js';
-import { resolveSocialMediaProvider } from '@/helpers/resolveSocialMediaProvider.js';
-import { useCurrentProfile } from '@/hooks/useCurrentProfile.js';
 import type { Profile } from '@/providers/types/SocialMedia.js';
-import { useGlobalState } from '@/store/useGlobalStore.js';
 
 interface InfoProps {
     profile: Profile;
 }
 
 export function Info({ profile }: InfoProps) {
-    const currentSource = useGlobalState.use.currentSource();
-    const currentSocialSource = narrowToSocialSource(currentSource);
-    const myProfile = useCurrentProfile(currentSocialSource);
-    const myProfileId = myProfile?.profileId;
     const profileId = profile.profileId;
 
     const source = profile.source;
-    const enabledMutuals = !isSameProfile(myProfile, profile);
-    // Fetch the first page with useInfiniteQuery, the same as
-    // MutualFollowersList, to make it reuseable in MutualFollowersList
-    const { data: mutuals } = useInfiniteQuery({
-        enabled: enabledMutuals,
-        queryKey: ['profiles', source, 'mutual-followers', myProfileId, profileId],
-        queryFn: async () => {
-            const provider = resolveSocialMediaProvider(source);
-            return provider.getMutualFollowers(profile.profileId);
-        },
-        initialPageParam: '',
-        getNextPageParam: (lastPage) => lastPage?.nextIndicator?.id,
-        select: (data) => data.pages.flatMap((page) => page?.data ?? EMPTY_LIST),
-    });
-    const mutualCount = mutuals?.length;
 
     const followingCount = profile.followingCount || 0;
     const followerCount = profile.followerCount || 0;
@@ -78,7 +53,7 @@ export function Info({ profile }: InfoProps) {
                         </TextOverflowTooltip>
                         <ProfileVerifyBadge className="flex flex-shrink-0 items-center space-x-1" profile={profile} />
                         <div className="ml-auto flex items-center gap-2">
-                            {showAction ? <ProfileAction profile={profile} /> : null}
+                            <NoSSR>{showAction ? <ProfileAction profile={profile} /> : null}</NoSSR>
                         </div>
                     </div>
                     <span className="text-medium text-secondary">@{profile.handle}</span>
@@ -112,16 +87,18 @@ export function Info({ profile }: InfoProps) {
                         <data value={followerCount}>
                             <span className="font-bold text-lightMain">{nFormatter(followerCount)} </span>
                             <span className="text-secondary">
-                                {plural(followerCount, {
+                                {/* {plural(followerCount, {
                                     one: 'Follower',
                                     other: 'Followers',
-                                })}
+                                })} */}
+                                <Plural value={followerCount} one="Follower" other="Followers" />
                             </span>
                         </data>
                     </Link>
                 </div>
             </div>
-            {enabledMutuals && mutualCount ? (
+            <Mutuals profile={profile} />
+            {/* {enabledMutuals && mutualCount ? (
                 <div className="break-word col-[1/3] mt-3 flex items-center gap-2 leading-[22px] hover:underline sm:col-[2/3]">
                     <AvatarGroup profiles={mutuals.slice(0, 3)} AvatarProps={{ size: 30 }} />
                     <Link
@@ -147,7 +124,7 @@ export function Info({ profile }: InfoProps) {
                         )}
                     </Link>
                 </div>
-            ) : null}
+            ) : null} */}
         </div>
     );
 }
