+++ src/hooks/useChannelPageContext.ts
@@ -1,3 +1,5 @@
+'use client';
+
 import { useState } from 'react';
 import { createContainer } from 'unstated-next';
