+++ src/services/getArticleCover.ts
@@ -1,5 +1,6 @@
 import urlcat from 'urlcat';
 
+import { SITE_URL } from '@/constants/index.js';
 import { fetchJSON } from '@/helpers/fetchJSON.js';
 import { type Article, ArticlePlatform } from '@/providers/types/Article.js';
 import type { ResponseJSON } from '@/types/index.js';
@@ -10,7 +11,7 @@ export async function getArticleCover(article: Article): Promise<string | null> 
     if (article.coverUrl) return article.coverUrl;
     if (article.platform === ArticlePlatform.Mirror && article.origin) {
         const payload = await fetchJSON<ResponseJSON<LinkDigested>>(
-            urlcat('/api/oembed', {
+            urlcat(SITE_URL, '/api/oembed', {
                 link: article.origin,
             }),
         );
